requirements.txt
torch
huggingface_hub
transformers
numpy
Pillow
bentoml>=1.2.0a0
